Extract latency logic
